LayoutTests/html5lib/webkit-runner-expected.txt

CONSOLE MESSAGE: line 3: PASS
CONSOLE MESSAGE: line 3: FOO<span>BAR</span>BAZ
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
resources/tests14.dat: PASS

resources/webkit01.dat: PASS

resources/doctype01.dat: 2, 4, 5, 8, 9, 10, 11, 12, 13, 17, 18, 19, 20, 21, 22, 24, 26, 27, 31, 33

Test 2 of 33 in resources/doctype01.dat failed. Input:
<!dOctYpE HtMl>Hello
Got:
| <!DOCTYPE HtMl>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 4 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE >
| <html>
|   <head>
|   <body>
|     "Hello"

Test 5 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE >Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE >
| <html>
|   <head>
|   <body>
|     "Hello"

Test 8 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato taco>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 9 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato taco "ddd>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 10 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato sYstEM>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 11 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato sYstEM    >Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 12 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE   potato       sYstEM  ggg>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 13 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato SYSTEM taco  >Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 17 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato SYSTEMtaco "ddd">Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 18 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato grass SYSTEM taco>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 19 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato pUbLIc>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 20 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato pUbLIc >Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 21 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato pUbLIcgoof>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 22 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC goof>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 24 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC 'go'of'>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 26 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE potato PUBLIC "W3C-//dfdf" SYSTEM ggg>Hello
Got:
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE potato>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 27 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN"
   "http://www.w3.org/TR/html4/strict.dtd">Hello
Got:
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "Hello"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "Hello"

Test 31 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE root-element [SYSTEM OR PUBLIC FPI] "uri" [ 
<!-- internal declarations -->
]>
Got:
| <html>
|   <head>
|   <body>
|     "
]>"
Expected:
| <!DOCTYPE root-element>
| <html>
|   <head>
|   <body>
|     "
]>"

Test 33 of 33 in resources/doctype01.dat failed. Input:
<!DOCTYPE HTML SYSTEM "http://www.w3.org/DTD/HTML4-strict.dtd"><body><b>Mine!</b></body>
Got:
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       "Mine!"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "Mine!"
resources/scriptdata01.dat: 6, 17, 18, 19, 21, 22, 23, 24, 26

Test 6 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script></scriptx>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "</scriptx>BAR"

Test 17 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt>'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt>'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt>'</script>BAR"

Test 18 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt> -'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> -'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> -'</script>BAR"

Test 19 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt> --'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> --'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> --'</script>BAR"

Test 21 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt> --!>'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> --!>'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> --!>'</script>BAR"

Test 22 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt> -- >'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> -- >'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt> -- >'</script>BAR"

Test 23 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt '</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt '"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt '</script>BAR"

Test 24 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt/'</script>BAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt/'"
|     "BAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt/'</script>BAR"

Test 26 of 26 in resources/scriptdata01.dat failed. Input:
FOO<script>'<!-- <sCrIpt/'</script>BAR</script>QUX
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt/'"
|     "BAR"
|     "QUX"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <script>
|       "'<!-- <sCrIpt/'</script>BAR"
|     "QUX"
resources/comments01.dat: 3, 6, 8, 9

Test 3 of 9 in resources/comments01.dat failed. Input:
FOO<!-- BAR --   >BAZ
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     "BAZ"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR --    -->
|     "BAZ"

Test 6 of 9 in resources/comments01.dat failed. Input:
FOO<!-- BAR -- <QUX> -- MUX -- >BAZ
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     " -- MUX -- >BAZ"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  BAR -- <QUX> -- MUX --  -->
|     "BAZ"

Test 8 of 9 in resources/comments01.dat failed. Input:
FOO<!--->BAZ
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     "BAZ"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"

Test 9 of 9 in resources/comments01.dat failed. Input:
FOO<!-->BAZ
Got:
| <html>
|   <head>
|   <body>
|     "FOO"
|     "BAZ"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO"
|     <!--  -->
|     "BAZ"
